requirements.txt
Flask==0.10.1
Flask-Bootstrap==3.3.5.7
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==1.7.0
Flask-Moment==0.5.1
Flask-SQLAlchemy==2.1
Flask-SSLify==0.1.5
Flask-Script==2.0.5
Flask-SocketIO==2.2
Flask-WTF==0.12
ForgeryPy==0.1
Jinja2==2.8
Mako==1.0.3
MarkupSafe==0.23
SQLAlchemy==1.0.11
SQLAlchemy-Utils==0.31.6
WTForms==2.1
WTForms-Components==0.10.0
Werkzeug==0.11.3
alembic==0.8.4
argparse==1.2.1
blinker==1.4
decorator==4.0.6
dominate==2.1.16
gunicorn==19.4.5
infinity==1.3
intervals==0.6.0
itsdangerous==0.24
phonenumbers==7.2.4
psycopg2==2.6.1
python-editor==0.5
python-engineio==0.9.0
python-firebase==1.2
python-socketio==1.2
requests==1.1.0
six==1.10.0
validators==0.10
visitor==0.1.2
wheel==0.24.0
wsgiref==0.1.2
